+++ source/DSCResources/MSFT_xDnsServerSetting/MSFT_xDnsServerSetting.psm1
@@ -6,7 +6,7 @@ Import-Module -Name $script:dnsServerDscCommonPath
 
 $script:localizedData = Get-LocalizedData -DefaultUICulture 'en-US'
 
-$properties = 'LocalNetPriority', 'AutoConfigFileZones', 'MaxCacheTTL', 'AddressAnswerLimit', 'UpdateOptions', 'DisableAutoReverseZones', 'StrictFileParsing', 'ForwardingTimeout', 'NoRecursion', 'ScavengingInterval', 'DisjointNets', 'Forwarders', 'DefaultAgingState', 'EnableDirectoryPartitions', 'XfrConnectTimeout', 'AllowUpdate', 'DsAvailable', 'BootMethod', 'LooseWildcarding', 'DsPollingInterval', 'BindSecondaries', 'LogLevel', 'AutoCacheUpdate', 'EnableDnsSec', 'EnableEDnsProbes', 'NameCheckFlag', 'EDnsCacheTimeout', 'SendPort', 'WriteAuthorityNS', 'IsSlave', 'RecursionTimeout', 'ListenAddresses', 'DsTombstoneInterval', 'RecursionRetry', 'RpcProtocol', 'SecureResponses', 'RoundRobin', 'ForwardDelegations', 'DefaultNoRefreshInterval', 'MaxNegativeCacheTTL', 'DefaultRefreshInterval'
+$properties = 'LocalNetPriority', 'AutoConfigFileZones', 'MaxCacheTTL', 'AddressAnswerLimit', 'UpdateOptions', 'DisableAutoReverseZones', 'StrictFileParsing', 'ForwardingTimeout', 'NoRecursion', 'DisjointNets', 'Forwarders', 'EnableDirectoryPartitions', 'XfrConnectTimeout', 'AllowUpdate', 'DsAvailable', 'BootMethod', 'LooseWildcarding', 'DsPollingInterval', 'BindSecondaries', 'LogLevel', 'AutoCacheUpdate', 'EnableDnsSec', 'EnableEDnsProbes', 'NameCheckFlag', 'EDnsCacheTimeout', 'SendPort', 'WriteAuthorityNS', 'IsSlave', 'RecursionTimeout', 'ListenAddresses', 'DsTombstoneInterval', 'RecursionRetry', 'RpcProtocol', 'SecureResponses', 'RoundRobin', 'ForwardDelegations', 'MaxNegativeCacheTTL'
 
 <#
     .SYNOPSIS
@@ -74,18 +74,6 @@ function Get-TargetResource
 
     .PARAMETER BootMethod
         Initialization method for the DNS Server.
-
-    .PARAMETER DefaultAgingState
-        Default ScavengingInterval value set for all Active Directory-integrated
-        zones created on this DNS Server.
-
-    .PARAMETER DefaultNoRefreshInterval
-        No-refresh interval, in hours, set for all Active Directory-integrated zones
-        created on this DNS Server.
-
-    .PARAMETER DefaultRefreshInterval
-        Refresh interval, in hours, set for all Active Directory-integrated zones
-        created on this DNS Server.
 
     .PARAMETER DisableAutoReverseZones
         Indicates whether the DNS Server automatically creates standard reverse look
@@ -177,10 +165,6 @@ function Get-TargetResource
     .PARAMETER RpcProtocol
         RPC protocol or protocols over which administrative RPC runs.
 
-    .PARAMETER ScavengingInterval
-        Interval, in hours, between two consecutive scavenging operations performed
-        by the DNS Server.
-
     .PARAMETER SecureResponses
         Indicates whether the DNS Server exclusively saves records of names in the
         same subtree as the server that provided them.
@@ -235,18 +219,6 @@ function Set-TargetResource
         [Parameter()]
         [uint32]
         $BootMethod,
-
-        [Parameter()]
-        [bool]
-        $DefaultAgingState,
-
-        [Parameter()]
-        [uint32]
-        $DefaultNoRefreshInterval,
-
-        [Parameter()]
-        [uint32]
-        $DefaultRefreshInterval,
 
         [Parameter()]
         [bool]
@@ -343,10 +315,6 @@ function Set-TargetResource
         [Parameter()]
         [int16]
         $RpcProtocol,
-
-        [Parameter()]
-        [uint32]
-        $ScavengingInterval,
 
         [Parameter()]
         [bool]
@@ -437,18 +405,6 @@ function Set-TargetResource
     .PARAMETER BootMethod
         Initialization method for the DNS Server.
 
-    .PARAMETER DefaultAgingState
-        Default ScavengingInterval value set for all Active Directory-integrated
-        zones created on this DNS Server.
-
-    .PARAMETER DefaultNoRefreshInterval
-        No-refresh interval, in hours, set for all Active Directory-integrated zones
-        created on this DNS Server.
-
-    .PARAMETER DefaultRefreshInterval
-        Refresh interval, in hours, set for all Active Directory-integrated zones
-        created on this DNS Server.
-
     .PARAMETER DisableAutoReverseZones
         Indicates whether the DNS Server automatically creates standard reverse look
         up zones.
@@ -539,10 +495,6 @@ function Set-TargetResource
     .PARAMETER RpcProtocol
         RPC protocol or protocols over which administrative RPC runs.
 
-    .PARAMETER ScavengingInterval
-        Interval, in hours, between two consecutive scavenging operations performed
-        by the DNS Server.
-
     .PARAMETER SecureResponses
         Indicates whether the DNS Server exclusively saves records of names in the
         same subtree as the server that provided them.
@@ -598,18 +550,6 @@ function Test-TargetResource
         [Parameter()]
         [uint32]
         $BootMethod,
-
-        [Parameter()]
-        [bool]
-        $DefaultAgingState,
-
-        [Parameter()]
-        [uint32]
-        $DefaultNoRefreshInterval,
-
-        [Parameter()]
-        [uint32]
-        $DefaultRefreshInterval,
 
         [Parameter()]
         [bool]
@@ -706,10 +646,6 @@ function Test-TargetResource
         [Parameter()]
         [int16]
         $RpcProtocol,
-
-        [Parameter()]
-        [uint32]
-        $ScavengingInterval,
 
         [Parameter()]
         [bool]